requirements.txt
reportlab
requests
pdfminer3k
python-docx
PyPDF2
PyMuPDF
matplotlib
fpdf
opencv-python
jieba

ChatTTS
IPython
omegaconf
vocos
vector_quantize_pytorch
transformers